backend/pyproject.toml
[project]
name = "aegisforensics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.14.0",
    "google-genai>=1.36.0",
    "pydantic>=2.11.8",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
